Make quantityXXX fields in UpdateRequest optional
This commit makes the fields `quantityMin`, `quantityMax`, and
`quantityInterval` in the class `UpdateRequest` optional.

If you want to update e.g. the `quantityMin` value, use something like
this:

```
request.setQuantityMin(Optional.of(10.0))
```

If you want to clear the `quantityMin` value, use this code:

```
request.setQuantityMin(Optional.empty())
```

If you do not call `setQuantityMin(...)` at all, its current value will
not be modified in an `Update` call to the server.

See #7